# backend/openapi/planning-api.yaml
openapi: 3.0.3
info:
  title: Timetable Todo Planning API
  version: 1.0.0
  description: >-
    Local-only planning API for the timetable todo application. All times are Asia/Seoul wall times
    on a 15-minute grid inside the 08:00~22:00 planning window. Every mutation carries the expected
    optimistic version and every response carries a request correlation ID.
servers:
  - url: http://127.0.0.1:8080
    description: Local development server (loopback only)

tags:
  - name: tasks
    description: Task content, completion and placement commands
  - name: planning
    description: Bounded weekly plan queries

paths:
  /api/v1/planning/weeks/{weekStart}:
    get:
      tags: [planning]
      summary: Read one Monday-anchored week plus the unscheduled backlog
      description: >-
        FR-001, FR-002, FR-005. weekStart must be a Monday; a non-Monday value is rejected with
        WEEK_START_INVALID rather than silently normalized. The response covers exactly
        [weekStart, weekStart + 7 days) and includes completed placements.
      operationId: getWeeklyPlan
      parameters:
        - name: weekStart
          in: path
          required: true
          schema:
            type: string
            format: date
          example: '2026-09-07'
      responses:
        '200':
          description: The bounded weekly plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyPlanView'
        '400':
          $ref: '#/components/responses/BadRequest'

  /api/v1/tasks:
    get:
      tags: [tasks]
      summary: List tasks with allowlisted filters and bounded paging
      description: FR-010, NFR-005. Page size is 1~100 with a default of 25 and stable ID tie-breaking.
      operationId: listTasks
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/TaskStatus'
        - name: scheduled
          in: query
          required: false
          schema:
            type: boolean
        - name: priority
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Priority'
        - name: sort
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/TaskSort'
        - name: direction
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/SortDirection'
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 25
      responses:
        '200':
          description: One bounded page of tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskPageView'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      tags: [tasks]
      summary: Create a task in the unscheduled backlog
      description: FR-003. A new task always starts as TODO with no placement.
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskContentRequest'
      responses:
        '201':
          description: The created task
          headers:
            Location:
              description: Absolute path of the created task
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskView'
        '400':
          $ref: '#/components/responses/BadRequest'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'

  /api/v1/tasks/{id}:
    get:
      tags: [tasks]
      summary: Read one task
      description: FR-004.
      operationId: getTask
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
      responses:
        '200':
          description: The requested task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskView'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [tasks]
      summary: Replace the full task content set
      description: >-
        FR-004, FR-007. title, priority and estimateMinutes are required; an omitted description or
        dueDate clears that value. Completion state and placement are owned by their own endpoints.
        A changed estimate keeps the placement start and recalculates its end, so the response can be
        a 409 conflict that stores nothing.
      operationId: updateTask
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: The updated task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskView'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
    delete:
      tags: [tasks]
      summary: Delete a task after explicit confirmation
      description: FR-004, BR-024. Both the expected version and an explicit confirmation are required.
      operationId: deleteTask
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
        - name: expectedVersion
          in: query
          required: true
          schema:
            type: integer
            format: int64
            minimum: 0
        - name: confirmed
          in: query
          required: true
          schema:
            type: boolean
      responses:
        '204':
          description: The task was deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /api/v1/tasks/{id}/schedule:
    put:
      tags: [tasks]
      summary: Place or move a task on the timetable
      description: >-
        FR-006, FR-007. The end time is always derived from the stored estimate; a client-supplied
        end is not accepted. An overlap returns 409 with the proposed slot, the conflicting slot and
        the next free candidate, and stores nothing.
      operationId: scheduleTask
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleTaskRequest'
      responses:
        '200':
          description: The placed task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskView'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
    delete:
      tags: [tasks]
      summary: Remove the placement and keep all content
      description: FR-008. Idempotent; an already unscheduled task creates no new audit event.
      operationId: unscheduleTask
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
        - name: expectedVersion
          in: query
          required: true
          schema:
            type: integer
            format: int64
            minimum: 0
      responses:
        '200':
          description: The unscheduled task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskView'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /api/v1/tasks/{id}/completion:
    put:
      tags: [tasks]
      summary: Set the desired completion state
      description: FR-009. Idempotent; repeating the current state creates no duplicate audit event.
      operationId: setTaskCompletion
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetCompletionRequest'
      responses:
        '200':
          description: The task in its new completion state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskView'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'

components:
  parameters:
    TaskIdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid

  responses:
    BadRequest:
      description: The request violated a validation or domain rule; nothing was stored
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: The resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: A placement overlap or a stale optimistic version; nothing was stored
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    UnsupportedMediaType:
      description: The request media type is not supported
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'

  schemas:
    Priority:
      type: string
      enum: [LOW, MEDIUM, HIGH]
    TaskStatus:
      type: string
      enum: [TODO, COMPLETED]
    TaskSort:
      type: string
      enum: [CREATED_AT, DUE_DATE, PRIORITY, TITLE]
      default: CREATED_AT
    SortDirection:
      type: string
      enum: [ASC, DESC]
      default: DESC
    ScheduleResolutionMode:
      type: string
      enum: [PROPOSE_ONLY, ACCEPT_CANDIDATE]
      default: PROPOSE_ONLY
      description: >-
        Informational client intent. Accepting a candidate re-runs the full conflict evaluation
        against current data exactly like a first placement.

    ScheduleView:
      type: object
      description: A half-open interval [startTime, endTime) on one date
      required: [date, startTime, endTime]
      properties:
        date:
          type: string
          format: date
        startTime:
          type: string
          pattern: '^([01][0-9]|2[0-3]):[0-5][0-9]$'
          example: '09:00'
        endTime:
          type: string
          pattern: '^([01][0-9]|2[0-3]):[0-5][0-9]$'
          example: '09:45'

    TaskView:
      type: object
      description: Transport representation of a task; absent optional fields are omitted, not null
      required: [id, title, priority, estimateMinutes, status, version, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          maxLength: 2000
        priority:
          $ref: '#/components/schemas/Priority'
        estimateMinutes:
          type: integer
          format: int32
          minimum: 15
          maximum: 840
        dueDate:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/TaskStatus'
        schedule:
          $ref: '#/components/schemas/ScheduleView'
        version:
          type: integer
          format: int64
          minimum: 0
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    TaskPageView:
      type: object
      required: [content, page, size, totalElements, totalPages]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/TaskView'
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32

    WeeklyPlanView:
      type: object
      required: [weekStart, weekEndExclusive, scheduled, backlog]
      properties:
        weekStart:
          type: string
          format: date
        weekEndExclusive:
          type: string
          format: date
        scheduled:
          type: array
          items:
            $ref: '#/components/schemas/TaskView'
        backlog:
          type: array
          items:
            $ref: '#/components/schemas/TaskView'

    TaskContentRequest:
      type: object
      required: [title, priority, estimateMinutes]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          maxLength: 2000
          nullable: true
        priority:
          $ref: '#/components/schemas/Priority'
        estimateMinutes:
          type: integer
          format: int32
          minimum: 15
          maximum: 840
          multipleOf: 15
        dueDate:
          type: string
          format: date
          nullable: true

    UpdateTaskRequest:
      type: object
      description: Full content replacement; an omitted description or dueDate clears that value
      required: [expectedVersion, title, priority, estimateMinutes]
      properties:
        expectedVersion:
          type: integer
          format: int64
          minimum: 0
        title:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          maxLength: 2000
          nullable: true
        priority:
          $ref: '#/components/schemas/Priority'
        estimateMinutes:
          type: integer
          format: int32
          minimum: 15
          maximum: 840
          multipleOf: 15
        dueDate:
          type: string
          format: date
          nullable: true

    ScheduleTaskRequest:
      type: object
      required: [date, startTime, expectedVersion]
      properties:
        date:
          type: string
          format: date
        startTime:
          type: string
          pattern: '^([01][0-9]|2[0-3]):(00|15|30|45)$'
          example: '09:00'
        expectedVersion:
          type: integer
          format: int64
          minimum: 0
        resolutionMode:
          $ref: '#/components/schemas/ScheduleResolutionMode'

    SetCompletionRequest:
      type: object
      required: [completed, expectedVersion]
      properties:
        completed:
          type: boolean
        expectedVersion:
          type: integer
          format: int64
          minimum: 0

    ApiError:
      type: object
      description: >-
        The single error representation. Only an allowlisted code, an authored message, the request
        correlation ID and allowlisted field names are returned. Rejected values, task content, SQL,
        stack traces and paths are never included. Fields that do not apply are omitted.
      required: [code, message, requestId]
      properties:
        code:
          type: string
          enum:
            - VALIDATION_FAILED
            - MALFORMED_REQUEST
            - DELETION_NOT_CONFIRMED
            - TASK_TITLE_INVALID
            - TASK_DESCRIPTION_TOO_LONG
            - TASK_ESTIMATE_INVALID
            - TASK_TIME_INVALID
            - SCHEDULE_ALIGNMENT_INVALID
            - SCHEDULE_OUT_OF_WINDOW
            - SCHEDULE_DURATION_INVALID
            - WEEK_START_INVALID
            - PAGE_INVALID
            - PAGE_SIZE_INVALID
            - BACKLOG_LIMIT_INVALID
            - TASK_NOT_FOUND
            - RESOURCE_NOT_FOUND
            - METHOD_NOT_ALLOWED
            - UNSUPPORTED_MEDIA_TYPE
            - SCHEDULE_CONFLICT
            - STALE_TASK
            - INTERNAL_ERROR
        message:
          type: string
        requestId:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
        currentVersion:
          type: integer
          format: int64
          description: Present on STALE_TASK; the persisted version the client must reload
        conflict:
          $ref: '#/components/schemas/ConflictView'

    FieldError:
      type: object
      required: [field, code]
      properties:
        field:
          type: string
        code:
          type: string
        message:
          type: string

    ConflictView:
      type: object
      description: Present on SCHEDULE_CONFLICT; carries times only, never another task's content
      required: [proposed, conflicting]
      properties:
        proposed:
          $ref: '#/components/schemas/ScheduleView'
        conflicting:
          $ref: '#/components/schemas/ScheduleView'
        nextCandidate:
          $ref: '#/components/schemas/ScheduleView'